pyproject.toml
[tool.poetry]
name = "talktodoc"
version = "0.1.0"
description = ""
authors = ["Aryan-401 <aryan.raj.garg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
langchain-groq = "^0.2.1"
langgraph = "^0.2.59"
python-dotenv = "^1.0.1"
groq = "^0.13.1"
pytest = "^8.3.4"
markitdown = "^0.0.1a3"
langchain-qdrant = "^0.2.0"
langchain-community = "^0.3.13"
streamlit = "^1.44.1"
nltk = "^3.9.1"
transformers = "^4.51.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
